Fixes webpack dev server error
Error encountered:
ERROR in chunk runtime [entry]
[name].[chunkhash].js
Cannot use [chunkhash] for chunk in '[name].[chunkhash].js' (use [hash]
instead)
